### 2022-08-22 #4696 @editorialbot
[PRE REVIEW]: WATTS: Workflow and template toolkit for simulation
**Submitting author:** @paulromano (<a href="http://orcid.org/0000-0002-1147-045X">Paul Romano</a>)
**Repository:** https://github.com/watts-dev/watts
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v0.3.0
