### 2022-08-24 #4696 @paulromano
@danielskatz Thanks for kicking this off. To respond to your question:

- Parsl is focused on enabling the execution of Python programs in parallel across many different types of systems, which is quite different than the aim of watts.
- Balsam is focused on managing batch job submission on clusters and does have a similar idea of allowing one to specify "parameters" that allow multiple jobs to execute for a given set of parameters.
- watts is primarily focused on managing templated input files for applications that are rendered using user-defined parameters. This is in contrast to Balsam where parameters are used as command-line options.

In principle, Balsam and watts could be used in conjunction with one another.
