### 2022-08-24 #4696 @danielskatz
thanks @paulromano - can you update the paper a bit to make this clear? (You don't need to mention these specific packages, but do clarify the role of watts in the workflow context compared to other kinds of workflow/scheduling systems)
