### 2022-08-24 #4696 @danielskatz
👋 @andrewryh - would you be willing to review this JOSS submission?
