### 2022-08-24 #4696 @danielskatz
👋 @gwenchee - would you be willing to review this JOSS submission?
