### 2022-08-24 #4696 @paulromano
> can you update the paper a bit to make this clear?

Absolutely -- I'll update the paper to expand on what we mean by "workflow" relative to other applications
