+++ paper.bib
@@ -184,3 +184,65 @@
   number = {NUREG/CR-5535-Vol.5},
   address = {Washington, DC},
 }
+
+@article{snakemake,
+  author = {Mölder, F and Jablonski, KP and Letcher, B and Hall, MB and Tomkins-Tinch, CH and Sochat, V and Forster, J and Lee, S and Twardziok, SO and Kanitz, A and Wilm, A and Holtgrewe, M and Rahmann, S and Nahnsen, S and Köster, J},
+  title = {Sustainable data analysis with Snakemake},
+  journal = {F1000Research},
+  year = 2021,
+  volume = 10,
+  number = 33,
+  doi = {10.12688/f1000research.29032.2},
+}
+
+@article{merlin,
+author = {J. Luc Peterson and Ben Bay and Joe Koning and Peter Robinson and Jessica Semler and Jeremy White and Rushil Anirudh and Kevin Athey and Peer-Timo Bremer and Francesco {Di Natale} and David Fox and Jim A. Gaffney and Sam A. Jacobs and Bhavya Kailkhura and Bogdan Kustowski and Steven Langer and Brian Spears and Jayaraman Thiagarajan and Brian {Van Essen} and Jae-Seung Yeom},
+  title = {Enabling machine learning-ready HPC ensembles with Merlin},
+  journal = {Future Generation Computer Systems},
+  volume = 131,
+  pages = {255--268},
+  year = 2022,
+  doi = {10.1016/j.future.2022.01.024},
+}
+
+@article{aiida,
+author = {Martin Uhrin and Sebastiaan P. Huber and Jusong Yu and Nicola Marzari and Giovanni Pizzi},
+  title = {Workflows in AiiDA: Engineering a high-throughput, event-based engine for robust and modular computational workflows},
+  journal = {Computational Materials Science},
+  volume = 187,
+  pages = {110086},
+  year = 2021,
+  doi = {10.1016/j.commatsci.2020.110086},
+}
+
+@article{scipipe,
+  author = {Lampa, Samuel and Dahlö, Martin and Alvarsson, Jonathan and Spjuth, Ola},
+  title = "{SciPipe: A workflow library for agile development of complex and dynamic bioinformatics pipelines}",
+  journal = {GigaScience},
+  volume = 8,
+  number = 5,
+  year = 2019,
+  month = apr,
+  doi = {10.1093/gigascience/giz044},
+}
+
+@inproceedings{parsl,
+  author = {Babuji, Yadu and Woodard, Anna and Li, Zhuozhao and Katz, Daniel S. and Clifford, Ben and Kumar, Rohan and Lacinski, Lukasz and Chard, Ryan and Wozniak, Justin M. and Foster, Ian and Wilde, Michael and Chard, Kyle},
+  title = {Parsl: Pervasive Parallel Programming in Python},
+  year = 2019,
+  publisher = {Association for Computing Machinery},
+  address = {New York, NY, USA},
+  doi = {10.1145/3307681.3325400},
+  booktitle = {Proceedings of the 28th International Symposium on High-Performance Parallel and Distributed Computing},
+  pages = {25--36},
+}
+
+
+@misc{balsam,
+  author = {Salim, Michael A. and Uram, Thomas D. and Childers, J. Taylor and Balaprakash, Prasanna and Vishwanath, Venkatram and Papka, Michael E.},
+  title = {Balsam: Automated Scheduling and Execution of Dynamic, Data-Intensive HPC Workflows},
+  publisher = {arXiv},
+  doi = {10.48550/arxiv.1909.08704},
+  url = {https://arxiv.org/abs/1909.08704},
+  year = {2019},
+}
