### 2022-08-29 be36f62
adds solver flag to settings template to match ABCE update
