+++ examples/1App_ABCE_C2N/watts_exec.py
@@ -22,7 +22,7 @@ params['run_ALEAF'] = "False"
 
 template_name = "abce_template.txt"
 results_path = Path.cwd() / 'results' / f"{n_steps}_periods_ALEAF_{params['run_ALEAF']}_99"
-results_path.mkdir(exist_ok=True)
+results_path.mkdir(exist_ok=True, parents=True)
 
 watts.Database.set_default_path(results_path)
 
