### 2022-09-02 #4735 @editorialbot
[REVIEW]: WATTS: Workflow and template toolkit for simulation
**Submitting author:** @paulromano (<a href="http://orcid.org/0000-0002-1147-045X">Paul Romano</a>)
**Repository:** https://github.com/watts-dev/watts
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v0.4.0
