### 2022-09-02 221a87f
consolidates self._executable in abce plugin
