### 2022-09-19 #4735 @sskutnik
Comment / suggestion on the paper; not well-formed enough to form a PR yet, but I wanted to put this out for discussion with the authors and other reviewers:

It seems like a substantial component of the "lift" being done by Watts specifically is in the [https://watts.readthedocs.io/en/latest/reference/generated/watts.Plugin.html#watts.Plugin](plugin class). That is, it acts as a functional adapter to the output generated by specific codes (like MCNP, MOOSE, OpenMC, etc.). The input generation largely seems like a function of Jinja (which is good and useful) with some supplemental physical unit conversion (also useful! but implemented outside this).

Assuming this is the case, I feel like the "plugin" interface as a core component of the Watts framework deserves a bit more attention / highlighting in the paper. The "what it does" part is fine, but from the perspective of a potential power user / contributor, part of what I would want to understand is what makes this all "tick"; how do I expand upon it to fit my code system of interest? Even just briefly going over the basic idea of the plugin in terms of what it's providing the framework (i.e., what's being implemented under the hood) would be useful to better contextualizing what's going on.

@munkm @yadudoc what do you think?
