### 2022-09-19 #4735 @sskutnik
Other comments; It would probably be a good idea to explicitly document required Python modules, beyond what is managed by Pip. Installation is easy and is readily reproduced from the instructions, but an explicit list of dependencies would be useful for users installing offline.

Repo has good documentation for developer contributions and use examples; I don't know how stringent we want to be with respect to "user support" documentation. Possibly this could be cleared up a little in the top-level package overview.

Otherwise, I think my one and only sticking point would be to flesh out a bit more how the plugin architecture works; with respect to WATTS, including some brief discussion of how this is bespoke to each package supported.
