### 2022-10-06 watts-dev/watts#61 @yadudoc
OpenMC required for tests
I'm over from JOSS looking at this repo to make sure everything checks out. I've looked over the docs quickly and everything looks really clean and understandable. I need to get the tests running on my laptop and I'm running into some issues that I'm hoping to get help with.

I've installed  `watts` from source, and when I run `pytest tests` I get the following (relevant snippet):
```
ImportError while importing test module '/tmp/watts/tests/test_plugin_openmc.py'.
Traceback:
/usr/lib/python3.8/importlib/__init__.py:127: in import_module
tests/test_plugin_openmc.py:7: in <module>
    import openmc
E   ModuleNotFoundError: No module named 'openmc'
```

I'm using conda, so I followed the install instructions here `https://docs.openmc.org/en/stable/quickinstall.html` and installed from mamba, and even still I get the same error. I've checked with `python3 -c "import openmc; print(openmc.__version__)"` to confirm that openmc is indeed installed. Any advice on this is much appreciated.

Linking [https://github.com/openjournals/joss-reviews/issues/4735]
