### 2022-10-06 watts-dev/watts#61 @paulromano
@yadudoc The only thing I can think is that when you call `pytest`, it's picking it up from a different Python environment, not the conda environment that has OpenMC installed in it. Perhaps try running `python3 -m pytest tests` instead?
