### 2022-10-06 watts-dev/watts#62 @yadudoc
Pointing to usage examples from docs
Following up on [https://github.com/openjournals/joss-reviews/issues/4735], I see that there is quite a collection of usage examples over at https://github.com/watts-dev/watts/tree/development/examples. Unfortunately, I really do not know have enough domain knowledge figure out where in those examples to start. I was hoping that perhaps a domain-specific "Hello world" example could be added straight to docs, perhaps on the `https://watts.readthedocs.io/en/latest/user/getting_started.html` page to get a novice user started.

I naively tried to run some scripts in those example dirs and it throws missing dir errors. It might be worthwhile to foolproof these scripts, but this probably isn't necessary to proceed with the JOSS review.
