### 2022-10-06 watts-dev/watts#63 @yadudoc
Add a note on support/help to the docs
This is related to the ongoing JOSS review [https://github.com/openjournals/joss-reviews/issues/4735].

It might be useful to have a tab on your docs to help guide novice users to the right place when seeking support. Perhaps point them to github issues, slack, or email, but having the preferred channels defined helps a lot.
