### 2022-10-06 #4735 @yadudoc
I've completed my review for the most part, and I've added three issues over on the github repo for watts detailing missing items. I'll come back and check off the missing boxes once the authors have had a chance to take a look.

Overall, I think this package provides values, the documentation is clean and understandable and the article does a good job of framing the problem. I think the plugin architecture for updating the application templates is quite nice, and I do wonder if the authors have considered integration with other python based workflow tools (as a workflow dev for parsl, this is cool!).
Another point to note is that "workflow" means many different things in different communities, it might be useful to define what it means to you. For, eg. is it possible to compose any DAG? Similarly, I only figured out what the execution context looks like after going over the "Detailed Usage" section in the documentation. As a user, I would like to know whether any task-parallelism is supported or whether the tasks run serially. The last couple of comments are purely from my curiosity and not necessary to move the review forward. Apologies to the authors on the delay in getting this review completed.

Pending issues from docs and tests:
1. https://github.com/watts-dev/watts/issues/63
2. https://github.com/watts-dev/watts/issues/62
3. https://github.com/watts-dev/watts/issues/61
