### 2022-10-06 watts-dev/watts#61 @yadudoc
@paulromano That definitely helped. I'm getting new errors now and the error message and troubleshooting guide over on OpenMC seems to indicate that I'm missing the `OPENMC_CROSS_SECTIONS` env var and an XML file.

Have you considered dropping a mock XML file into tests and setting the required env var from conftests?
Here's a snippet of the error:

```
tests/test_plugin_openmc.py FFF
...
                error_msg = 'OpenMC aborted unexpectedly.'
            error_msg = ' '.join(error_msg.split())

>           raise RuntimeError(error_msg)
E           RuntimeError: No cross_sections.xml file was specified in materials.xml or in the OPENMC_CROSS_SECTIONS environment variable. OpenMC needs such a file to identify where to find data libraries. Please consult the user's guide at https://docs.openmc.org/ for information on how to set up data libraries.

/home/yadu/anaconda3/envs/watts/lib/python3.10/site-packages/openmc/executor.py:116: RuntimeError

```
