### 2022-10-14 watts-dev/watts#64 @munkm
Add clarification of input file deletion from Database
This is for my review for https://github.com/openjournals/joss-reviews/issues/4735  .

In the documentation for the detailed usage of WATTS on the Database class, it is mentioned that if one uses the `Database.clear()` method that
> Be aware that clearing the database will delete all the corresponding results on disk, including input and output files from the workflow.

I am curious if this includes deleting the input files where they were specified on disc, or just the paths to those locations in the database? Does this also include the `extra_inputs` and `extra_template_inputs` files if they are specified?

I also think it would be useful to show how to clear out single objects from the database, rather than the whole database. I can imagine a use case where a user might want to delete a single object that was an error and not all of the entries in the database.
