### 2022-10-14 watts-dev/watts#65 @munkm
Specific developer requirements for a new `Plugin` 
This is for my review of https://github.com/openjournals/joss-reviews/issues/4735

This is a non-blocking issue for that review! Thank you for this package and I can certainly see the use of this package (especially with the examples shown in the documentation!).

I can imagine that users of your package will be excited to contribute new plugins, so I think it would be helpful to describe in a specific section of the developer guide how to contribute a new plugin. What is minimally required for a new plugin, and do you recommend a specific plugin as an example for them to get started?
