### 2022-10-14 #4735 @munkm
Hi!!! I am really excited about this package and thank you for your contribution to JOSS. I opened two issues that I think will help clarify things in the documentation for your future users, but watts-dev/watts#65 is not blocking for my review. However, I do think it's related and complementary to the original comment made by @sskutnik about power users.  I have finished my review but I have two questions on your references:

- Are the MNCP 6.2 release notes the preferred reference for MCNP? I am not totally sure based on their website, but it seems for 6.3 their main reference is the user's manual. https://mcnp.lanl.gov/reference_collection.html#mcnp6_refs
- ABCE is a plugin mentioned in the docs, but I don't see it referenced in the paper. Is there no reference? Perhaps you could add a sentence describing it too.
