### 2022-10-31 c914624
Change OpenMC plugin test to use a local data library
