### 2022-10-31 948130a
Add Database.remove method to remove a single result
