### 2022-11-01 6a59a73
Initial implementation of generic plugin
