+++ src/watts/__init__.py
@@ -11,6 +11,7 @@ from .plugin_mcnp import *
 from .plugin_serpent import *
 from .plugin_abce import *
 from .plugin_dakota import *
+from .results import *
 from .template import *
 from .parameters import *
 from .database import *
