+++ doc/source/reference/index.rst
@@ -11,6 +11,7 @@ API Reference
    watts.Database
    watts.Parameters
    watts.Plugin
+   watts.PluginGeneric
    watts.PluginABCE
    watts.PluginMCNP
    watts.PluginMOOSE
