### 2022-11-01 06046ca
Merge PluginGeneric into TemplatePlugin
