+++ src/watts/plugin_pyarc.py
@@ -10,7 +10,7 @@ from typing import Mapping, List, Optional
 
 from .fileutils import PathLike
 from .parameters import Parameters
-from .plugin import TemplatePlugin
+from .plugin import PluginGeneric
 from .results import Results
 
 
@@ -40,7 +40,7 @@ class ResultsPyARC(Results):
         self.results_data = results_data
 
 
-class PluginPyARC(TemplatePlugin):
+class PluginPyARC(PluginGeneric):
     """Plugin for running PyARC
 
     Parameters
@@ -67,12 +67,12 @@ class PluginPyARC(TemplatePlugin):
                  extra_inputs: Optional[List[str]] = None,
                  extra_template_inputs: Optional[List[PathLike]] = None,
                  show_stdout: bool = False, show_stderr: bool = False):
-        super().__init__(template_file, extra_inputs, extra_template_inputs,
-                         show_stdout, show_stderr)
-        self._executable = Path(os.environ.get('PyARC_DIR', 'PyARC.py'))
+        executable = Path(os.environ.get('PyARC_DIR', 'PyARC.py'))
+        super().__init__(executable, None, template_file, extra_inputs,
+                         extra_template_inputs, show_stdout, show_stderr)
         self.input_name = "pyarc_input.son"
 
-    @TemplatePlugin.executable.setter
+    @PluginGeneric.executable.setter
     def executable(self, exe: PathLike):
         if Path(exe).exists():
             raise RuntimeError(f"{self.plugin_name} executable '{exe}' is missing.")
