### 2022-11-01 8b54406
Add developer's guide documentation on plugins
