### 2022-11-01 ff7de66
Add a complete example in the getting started section
