### 2022-11-02 aa81b8a
Allow execute_command to be a single str
