### 2022-11-02 28aa588
Fix setting PyARC executable property
