### 2022-11-02 d7ed191
Fix UnboundLocalError in plugin_sas.py
