### 2022-11-03 #4735 @paulromano
Hi @danielskatz There is indeed. I've had a few PRs to the watts repo addressing the reviewer feedback. Most of these have already been merged and I'm just waiting on one more (watts-dev/watts#71). I've also made some local updates to the text of the paper based on these updates and will push that once the review of the aforementioned PR is complete. I'll add another comment here once I believe all the feedback has been addressed and it's ready for another look.

Thanks for checking in!
