### 2022-11-04 14f0424
Allow executable to be specified in PluginMOOSE constructor
