+++ examples/MultiStep_Griffin-BISON-Sockeye_MR/watts_exec.py
@@ -1,6 +1,7 @@
 # SPDX-FileCopyrightText: 2022 UChicago Argonne, LLC
 # SPDX-License-Identifier: MIT
 
+from pathlib import Path
 import os
 import watts
 
@@ -11,15 +12,15 @@ Second step is a Null transient simulation to confirm the steady state obtained
 Third step is a real transient simulation induced by loss of cooling capacity
 """
 
-# set your SuperMoose directory as SUPER_MOOSE
+# set your SuperMoose directory as SUPER_MOOSE_DIR
 # As Sockeye is run through dynamic linking
 # set your Sockeye directory as SOCKEYE_DIR
 
 if not os.getenv("SOCKEYE_DIR"):
     raise RuntimeError("SOCKEYE_DIR must be set to enable this example.")
 # MOOSE app type to run
-moose_app_type = "super_moose"
-app_dir = os.environ[moose_app_type.upper() + "_DIR"]
+app_dir = Path(os.enrivon["SUPER_MOOSE_DIR"])
+moose_exec = app_dir / "super_moose-opt"
 
 # Steady State Parameters
 params_ss = watts.Parameters()
@@ -42,8 +43,11 @@ params_ss.show_summary(show_metadata=False, sort_by='key')
 # MOOSE Workflow for steady state
 print("Steady-state calculation")
 mpi_args = ['mpiexec', '-n', '40']
-moose_plugin_ss = watts.PluginMOOSE('MP_ss_griffin.tmpl', extra_inputs=['MP_ss_moose.i', 'MP_ss_sockeye.i', '3D_unit_cell_FY21_level-1_bison.e', '3D_unit_cell_FY21_supersimple.e', 'unitcell_nogap_hom_xml_G11_df_MP.xml'])
-moose_plugin_ss.executable = app_dir + "/" + moose_app_type.lower() + "-opt"
+moose_plugin_ss = watts.PluginMOOSE(
+    'MP_ss_griffin.tmpl',
+    executable=moose_exec,
+    extra_inputs=['MP_ss_moose.i', 'MP_ss_sockeye.i', '3D_unit_cell_FY21_level-1_bison.e', '3D_unit_cell_FY21_supersimple.e', 'unitcell_nogap_hom_xml_G11_df_MP.xml']
+)
 moose_result_ss = moose_plugin_ss(params_ss, mpi_args=mpi_args)
 for key in moose_result_ss.csv_data:
     print(key, moose_result_ss.csv_data[key])
@@ -58,8 +62,12 @@ params_trN = params_ss
 
 # MOOSE Workflow for Null transient
 print("Null transient calculation")
-moose_plugin_trN = watts.PluginMOOSE('MP_trN_griffin.tmpl', show_stdout=False, extra_inputs=['MP_trN_moose.i', 'MP_trN_sockeye.i', 'unitcell_nogap_hom_xml_G11_df_MP.xml'])
-moose_plugin_trN.executable = app_dir + "/" + moose_app_type.lower() + "-opt"
+moose_plugin_trN = watts.PluginMOOSE(
+    'MP_trN_griffin.tmpl',
+    executable=moose_exec,
+    show_stdout=False,
+    extra_inputs=['MP_trN_moose.i', 'MP_trN_sockeye.i', 'unitcell_nogap_hom_xml_G11_df_MP.xml']
+)
 moose_result_trN = moose_plugin_trN(params_trN, mpi_args=mpi_args)
 for key in moose_result_trN.csv_data:
     print(key, moose_result_trN.csv_data[key])
@@ -78,8 +86,12 @@ params_tr = params_ss
 
 # MOOSE Workflow for transient
[…]
