+++ examples/1App_MOOSE-MultiApp_Simple/watts_exec.py
@@ -2,11 +2,13 @@
 # SPDX-License-Identifier: MIT
 
 """
-This example provides a demonstration on how to use WATTS to perform a simple simulation leveraging MOOSE's MultiApps system.
+This example provides a demonstration on how to use WATTS to perform a simple
+simulation leveraging MOOSE's MultiApps system.
 """
 
-from math import cos, pi
+from pathlib import Path
 import os
+
 import watts
 
 params = watts.Parameters()
@@ -18,12 +20,14 @@ params['Swelling_Coefficient'] = 3000.0 # K.
 params.show_summary(show_metadata=False, sort_by='key')
 
 # MOOSE Workflow
-# set your BISON directorate as BISON_DIR
+# set your BISON directory as BISON_DIR
 
-moose_app_type = "bison"
-app_dir = os.environ[moose_app_type.upper() + "_DIR"]
-moose_plugin = watts.PluginMOOSE('main.tmpl', extra_inputs=['main_in.e', 'sub.i'])
-moose_plugin.executable = app_dir + "/" + moose_app_type.lower() + "-opt"
+app_dir = Path(os.environ["BISON_DIR"])
+moose_plugin = watts.PluginMOOSE(
+    'main.tmpl',
+    executable=app_dir / 'bison-opt',
+    extra_inputs=['main_in.e', 'sub.i']
+)
 moose_result = moose_plugin(params)
 for key in moose_result.csv_data:
     print(key, moose_result.csv_data[key])
