### 2022-11-04 #4735 @paulromano
Alright, the final PR has been merged and I've made corresponding updates in the `joss-paper` branch of watts. To respond specifically to a few of the comments made here:

> Assuming this is the case, I feel like the "plugin" interface as a core component of the Watts framework deserves a bit more attention / highlighting in the paper. The "what it does" part is fine, but from the perspective of a potential power user / contributor, part of what I would want to understand is what makes this all "tick"; how do I expand upon it to fit my code system of interest? Even just briefly going over the basic idea of the plugin in terms of what it's providing the framework (i.e., what's being implemented under the hood) would be useful to better contextualizing what's going on.

@sskutnik This is a really good point. I've expanded the description of the `Plugin` classes in the paper to describe in more detail how they fit in to the overall system. I've also added a new [section](https://watts.readthedocs.io/en/latest/dev/plugins.html) in our developer's guide that goes over how to connect new codes into watts.

> Another point to note is that "workflow" means many different things in different communities, it might be useful to define what it means to you. For, eg. is it possible to compose any DAG? Similarly, I only figured out what the execution context looks like after going over the "Detailed Usage" section in the documentation. As a user, I would like to know whether any task-parallelism is supported or whether the tasks run serially.

@yadudoc This is discussed a bit in the paper itself, namely the relationship of `watts` to other workflow systems and specifically what `watts` does not do. It's a really interesting question though and I'm planning on spending some time after this to explore how `watts` and other workflow systems like Parsl could fit together! Thanks for all your helpful comments and suggestions :pray:

> * Are the MNCP 6.2 release notes the preferred reference for MCNP? I am not totally sure based on their website, but it seems for 6.3 their main reference is the user's manual. https://mcnp.lanl.gov/reference_collection.html#mcnp6_refs
> * ABCE is a plugin mentioned in the docs, but I don't see it referenced in the paper. Is there no reference? Perhaps you could add a sentence describing it too.

@munkm Thanks for pointing out the MCNP 6.3 release notes. Looks like those are hot off the press; I've updated the paper to cite those instead of the older 6.2 notes.

Regarding ABCE, unfortunately there is no reference at present. If it's OK with you, we can add a reference 1) in the documentation of `watts` once a reference becomes available and 2) in a future update of the paper for a new release.

Once again, thanks to all for the excellent feedback and helping to improve the software further!
