### 2022-11-07 #4735 @danielskatz
👋 @paulromano - At this point could you:

I can then move forward with testing acceptance of the submission, then proofreading the final draft.
