+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "flit_core.buildapi"
 
 [project]
 name = "watts"
-version = "0.4.0"
+version = "0.4.1-dev"
 authors = [
     { name = "UChicago Argonne, LLC", email = "watts@anl.gov" },
 ]
