### 2022-11-10 #4735 @paulromano
@danielskatz I've tagged a new release, [v0.4.0](https://github.com/watts-dev/watts/releases/tag/v0.4.0), and it's been archived at:
[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.7309095.svg)](https://doi.org/10.5281/zenodo.7309095)
