### 2022-11-10 #4735 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
